Change macro goal target types from Float to Int and update database schema
- Update `protein`, `carbs`, and `fat` fields in `MacroGoal` and `MacroGoalEntity` models from `Float` to `Int`.
- Add Room database migration (v7 to v8) to reconstruct the `macro_goals` table with integer macro columns.
- Update `SaveMacroGoalUseCase` and `MacroViewModel` to parse and store macro goal targets as integers.
- Adjust `HistoryScreen` target formatting to convert integer goals for progress calculations and omit unnecessary trailing decimals in the UI.
- Update unit tests in `MacroViewModelTest` and `HistoryViewModelTest` to reflect the new integer macro goal model.

+++ apps/nutrition-tracker/feature-api/src/main/java/com/personalapps/suite/nutrition/feature/api/model/MacroGoal.kt
@@ -3,7 +3,7 @@ package com.personalapps.suite.nutrition.feature.api.model
 data class MacroGoal(
     val id: Int = 1,
     val calories: Int,
-    val protein: Float,
-    val carbs: Float,
-    val fat: Float
+    val protein: Int,
+    val carbs: Int,
+    val fat: Int
 )
